+++ openapi_client/models/__init__.py
@@ -4,7 +4,7 @@
 """
     3Di API
 
-    3Di simulation API (latest version: 3.0)   Framework release: 1.0.11   3Di core release: 2.0.10  deployed on:  11:20AM (UTC) on July 10, 2020  # noqa: E501
+    3Di simulation API (latest version: 3.0)   Framework release: 1.0.11   3Di core release: 2.0.12  deployed on:  13:48AM (UTC) on July 15, 2020  # noqa: E501
 
     The version of the OpenAPI document: 3.0
     Contact: info@nelen-schuurmans.nl
@@ -121,6 +121,8 @@ from openapi_client.models.post_processing_status import PostProcessingStatus
 from openapi_client.models.potential_breach import PotentialBreach
 from openapi_client.models.profile import Profile
 from openapi_client.models.progress import Progress
+from openapi_client.models.pump_discharge_graph import PumpDischargeGraph
+from openapi_client.models.pump_discharge_graph_request import PumpDischargeGraphRequest
 from openapi_client.models.raster import Raster
 from openapi_client.models.raster_edit import RasterEdit
 from openapi_client.models.raster_edit_urls import RasterEditUrls